Move `ErrorBoundary` and `QueryCacheProvider` into `src/components`.
